test: IFS-2400 move AutoConfiguration specific Testcase into IsySecurityAutoConfigurationTest
